requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
lap==0.5.13
MarkupSafe==3.0.2
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.13.0.92
packaging==26.0
pillow==12.0.0
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
torch==2.5.1+cu121
torchvision==0.20.1+cu121
triton==3.1.0
typing_extensions==4.15.0
ultralytics==8.4.17
ultralytics-thop==2.0.18
urllib3==2.6.3
